app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
google-generativeai==0.8.4
httpx==0.27.0
selenium==4.18.1
webdriver-manager==4.0.1

# pip install -r requirements.txt